Fix ../src/libslic3r/QuadricEdgeCollapse.cpp:565:22: warning: comparison of integer expressions of different signedness: 'const int' and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare] ../src/slic3r/GUI/Gizmos/GLGizmoSimplify.cpp:48:15: warning: unused variable 'min_triangle_count' [-Wunused-variable] ../src/slic3r/GUI/Gizmos/GLGizmoSimplify.cpp:160:9: warning: unused variable 'wanted_count' [-Wunused-variable] ..\src\slic3r\GUI\Gizmos\GLGizmoSimplify.cpp(167): warning C4305: '=': truncation from 'double' to 'float'

+++ src/libslic3r/QuadricEdgeCollapse.cpp
@@ -562,7 +562,7 @@ bool QuadricEdgeCollapse::degenerate(uint32_t          vi,
         if (e_info.t_index == ti1) continue; // ti1 will be deleted
         const Triangle &t = indices[e_info.t_index];
         for (size_t i = 0; i < 3; ++i)
-            if (t[i] == vi) return true;
+            if (static_cast<uint32_t>(t[i]) == vi) return true;
     }
     return false;
 }
